fixing .gitignore and dev-stop.sh

+++ dev-stop.sh
@@ -1,3 +1,5 @@
 #!/bin/bash
-kill -15 $(cat chess.pid)
-rm chess.pid
+if [ -f chess.pid ]; then
+  kill -15 $(cat chess.pid)
+  rm chess.pid
+fi
